# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API service for Oil Bank System",
    "title": "Oil Bank System Backend API Service",
    "contact": {},
    "version": "0.0.1"
  },
  "basePath": "/v1",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format: Bearer {token}"
    }
  },
  "paths": {
...
    "/users": {
      "post": {
        "description": "Register a new user account",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Register a new user",
        "parameters": [
          {
            "description": "User registration details",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/HTTPResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/User"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "409": {
            "description": "Conflict - User already exists",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Get user by ID",
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Get user details",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/HTTPResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/User"
                    }
                  }
                }
              ]
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          }
        }
      },
      "patch": {
        "description": "Update user information",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Update user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "User update details",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/HTTPResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/User"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete user by ID",
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Delete user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/HTTPResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "boolean"
                    }
                  }
                }
              ]
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          }
        }
      }
    },
    "/users/auth/login": {
      "post": {
        "description": "Authenticate user and receive access token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "User login",
        "parameters": [
          {
            "description": "Login credentials",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/HTTPResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/User"
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "404": {
            "description": "Not Found - User not found",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/HTTPErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserCreateRequest": {
      "type": "object",
      "required": ["username", "email", "password", "user_type"],
      "properties": {
        "username": {
          "type": "string",
          "minLength": 7,
          "example": "johndoe"
        },
        "email": {
          "type": "string",
          "format": "email",
          "example": "john@example.com"
        },
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "password123"
        },
        "user_type": {
          "type": "string",
          "enum": ["SELLER", "COLLECTOR", "COMPANY"],
          "example": "SELLER"
        }
      }
    },
    "UserLoginRequest": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "john@example.com"
        },
        "password": {
          "type": "string",
          "example": "password123"
        }
      }
    },
    "UserUpdateRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "johndoe_updated"
        },
        "email": {
          "type": "string",
          "format": "email",
          "example": "john.updated@example.com"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "user_type": {
          "type": "string",
          "enum": ["SELLER", "COLLECTOR", "COMPANY"]
        },
        "address": {
          "$ref": "#/definitions/UserAddress"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "access_token": {
          "type": "string",
          "description": "JWT access token (only returned on login)"
        }
      }
    },
    "UserAddress": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "street_address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "regency": {
          "type": "string"
        },
        "province": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "HTTPResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object"
        },
        "error": {
          "type": "object"
        }
      }
    },
    "HTTPErrorResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "nullable": true
        },
        "error": {
          "$ref": "#/definitions/HTTPError"
        }
      }
    },
    "HTTPError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 404
        },
        "message": {
          "type": "string",
          "example": "User not found"
        },
        "known_error": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}